Removing some debug code in the utils files Moving datasets to their own modules
